# api-specs.yaml
openapi: 3.0.0
info:
  title: 'Kenisty API'
  description: 'The main entry point for Kenisty project.'
  contact:
    name: 'Pola Eskandar'
    url: 'https://codingstreamer.com'
    email: eskandar.pola@codingstreamer.com
  license:
    name: MIT
  version: 1.0.0
paths:
  /v1/auth/register:
    post:
      operationId: 557dcb2ee3dfdae1e150def5ea1f831c
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterReadV1'
      responses:
        '201':
          description: 'Successfully logged in an existing user.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegisterWriteV1'
        '422':
          description: 'Failed validation.'
components:
  schemas:
    UserRegisterReadV1:
      title: UserRegisterReadV1
      properties:
        firstname:
          title: Firstname
          type: string
          maxLength: 255
          minLength: 3
          nullable: false
        lastname:
          title: Lastname
          type: string
          maxLength: 255
          minLength: 3
          nullable: false
        email:
          title: Email
          type: string
          maxLength: 255
          minLength: 3
          nullable: false
        password:
          title: Password
          type: string
          maxLength: 255
          minLength: 8
          nullable: false
        password_confirmation:
          title: 'Password Confirmation'
          type: string
          maxLength: 255
          minLength: 8
          nullable: false
      type: object
    ServerExceptionWriteV1:
      title: ServerExceptionWriteV1
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
    UserRegisterWriteV1:
      title: UserRegisterWriteV1
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserDataV1'
      type: object
    UserDataV1:
      title: UserDataV1
      properties:
        token:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
      type: object